requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
appnope==0.1.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
backcall==0.2.0
certifi==2020.12.5
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
decorator==4.4.2
discord-flags==2.1.1
discord.py==1.6.0
distlib==0.3.1
filelock==3.0.12
google-trans-new==1.1.9
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
identify==2.2.1
idna==2.10
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
isbnlib==3.10.6
isbntools==4.3.25
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
lxml==4.6.3
multidict==5.1.0
nodeenv==1.5.0
packaging==20.9
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
pre-commit==2.11.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
toml==0.10.2
tornado==6.1
tox==3.23.0
traitlets==5.0.5
translate==3.5.0
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.3
wcwidth==0.2.5
yarl==1.6.3
